requirements.txt
playwright
apache-airflow